SpectralKit

This is a very simple package for building blocks of spectral methods. Its intended audience is users who are familiar with the theory and practice of these methods, and prefer to assemble their code from modular building blocks. If you need an introduction, a book like Boyd (2001): Chebyshev and Fourier spectral methods is a good place to start.

The package is optimized for solving functional equations, as usually encountered in economics when solving discrete and continuous-time problems. It uses static arrays extensively to avoid allocation and unroll some loops. Key functionality includes evaluating a set of basis functions, their linear combination at arbitrary points in a fast manner, for use in threaded code. These should work seamlessly with automatic differentiation frameworks when derivatives are needed.

Introduction

In this package,

A basis is a finite family of functions for approximating other functions. The dimension of a basis tells you how many functions are in there, while domain can be used to query its domain.

A grid is vector of suggested gridpoints for evaluating the function to be approximated that has useful theoretical properties. You can contruct a collocation_matrix using this grid. Grids are associated with bases at the time of their construction: a basis with the same set of functions can have different grids.

basis_at returns an iterator for evaluating basis functions at an arbitrary point inside their domain. This iterator is meant to be heavily optimized and non-allocating. linear_combination is a convenience wrapper for obtaining a linear combination of basis functions at a given point.

A basis is constructed using

a primitive family on a fixed domain, eg Chebyshev,

a grid specification like InteriorGrid,

a number of parameters that determine the number of basis functions.

transformation(s) that project the domain of a primitive family into the domain.

The following example provides an overview of the interface for univariate function families:

julia> using SpectralKit
julia> basis = univariate_basis(Chebyshev, EndpointGrid(), # 5 Chebyshev polynomials on [0,2]
                                5, BoundedLinear(0, 2))Chebyshev polynomials (1st kind), grid w/ endpoints, dimension: 5
  domain (0.0,2.0) [linear transformation]
julia> is_function_basis(basis) # tells us that we support the interface belowtrue
julia> dimension(basis) # number of basis functions5
julia> domain(basis)            # domain(0.0, 2.0)
julia> grid(basis) # Gauss-Lobatto grid5-element Vector{Float64}:
 0.0
 0.2928932188134524
 1.0
 1.7071067811865475
 2.0
julia> collect(basis_at(basis, 0.41)) # iterator for basis functions at 0.415-element Vector{Float64}:
  1.0
 -0.5900000000000001
 -0.30379999999999985
  0.9484839999999999
 -0.8154111200000003
julia> θ = [1, 0.5, 0.2, 0.3, 0.001] # a vector of coefficients5-element Vector{Float64}:
 1.0
 0.5
 0.2
 0.3
 0.001
julia> linear_combination(basis, θ, 0.41) # combination at some value0.9279697888799999
julia> linear_combination(basis, θ)(0.41) # also as a callable0.9279697888799999

Constructing bases

Primitive families

Currently, only Chebyshev polynomials are implemented.

SpectralKit.Chebyshev — Type

struct Chebyshev{K} <: SpectralKit.FunctionBasis

The first N Chebyhev polynomials of the first kind, defined on [-1,1].

Grid specifications

SpectralKit.InteriorGrid — Type

struct InteriorGrid <: SpectralKit.AbstractGrid

Grid with interior points (eg Gauss-Chebyshev).

SpectralKit.EndpointGrid — Type

struct EndpointGrid <: SpectralKit.AbstractGrid

Grid that includes endpoints (eg Gauss-Lobatto).

Transformations

SpectralKit.BoundedLinear — Type

struct BoundedLinear{T<:Real} <: SpectralKit.UnivariateTransformation

Transform x ∈ (-1,1) to y ∈ (a, b), using $y = x ⋅ s + m$.

m and s are calculated and checked by the constructor; a < b is enforced.

SpectralKit.InfRational — Type

InfRational(A, L)

Chebyshev polynomials transformed to the domain (-Inf, Inf) using $y = A + L ⋅ x / √(1 - x^2)$, with L > 0.

0 is mapped to A.

SpectralKit.SemiInfRational — Type

SemiInfRational(A, L)

[-1,1] transformed to the domain [A, Inf) (when L > 0) or (-Inf,A] (when L < 0) using $y = A + L ⋅ (1 + x) / (1 - x)$.

When used with Chebyshev polynomials, also known as a “rational Chebyshev” basis.

Univariate bases

SpectralKit.univariate_basis — Function

univariate_basis(univariate_family, grid_kind, N, transformation)

Create a univariate basis from univariate_family, with the specified grid_kind and dimension N, transforming the domain with transformation.

parent is a univariate basis, transformation is a univariate transformation (supporting the interface described by UnivariateTransformation, but not necessarily a subtype). Univariate bases support SpectralKit.gridpoint.

Example

The following is a basis with 10 transformed Chebyshev polynomials of the first kind on $(3,∞)$, with equal amounts of nodes on both sides of 7 = 3 + 4 and an interior grid:

julia> basis = univariate_basis(Chebyshev, InteriorGrid(), 10, SemiInfRational(3.0, 4.0))
Chebyshev polynomials (1st kind), interior grid, dimension: 10
  domain (3.0,∞) [rational transformation with scale 4.0]

julia> dimension(basis)
10

julia> domain(basis)
(3.0, Inf)

Multivariate bases

SpectralKit.SmolyakParameters — Type

SmolyakParameters(B)
SmolyakParameters(B, M)

Parameters for Smolyak grids that are independent of the dimension of the domain.

Polynomials are organized into blocks of 1, 2, 2, 4, 8, 16, … polynomials (and corresponding gridpoints), indexed with a block index b that starts at 0. B ≥ ∑ bᵢ and 0 ≤ bᵢ ≤ M constrain the number of blocks along each dimension i.

SpectralKit.smolyak_basis — Function

smolyak_basis(univariate_family, grid_kind, smolyak_parameters, transformations)

Create a sparse Smolyak basis using univariate_family (eg Chebyshev), which takes grid_kind and a dimension parameter. B > 0 caps the sum of blocks, while M > 0 caps blocks along each dimension separately, and transformations is a tuple of transformations applied coordinate-wise.

Example

julia> basis = smolyak_basis(Chebyshev, InteriorGrid(), SmolyakParameters(3),
                             (BoundedLinear(2, 3), SemiInfRational(3.0, 4.0)))
Sparse multivariate basis on ℝ^2
  Smolyak indexing, ∑bᵢ ≤ 3, all bᵢ ≤ 3, dimension 29
  using Chebyshev polynomials (1st kind), interior grid, dimension: 9
  domain transformations
    (2.0,3.0) [linear transformation]
    (3.0,∞) [rational transformation with scale 4.0]

julia> dimension(basis)
29

julia> domain(basis)
((2.0, 3.0), (3.0, Inf))

Using bases

Introspection

SpectralKit.is_function_basis — Function

is_function_basis(F)

is_function_basis(f::F)

Test if the argument is a function basis, supporting the following interface:

domain for querying the domain,

dimension for the dimension,

basis_at for function evaluation,

grid to obtain collocation points.

linear_combination and collocation_matrix are also supported, building on the above.

Can be used on both types (preferred) and values (for convenience).

SpectralKit.dimension — Function

dimension(basis)

Return the dimension of basis, a positive Int.

SpectralKit.domain — Function

domain(basis)

The domain of a function basis. A tuple of numbers (of arbitrary type, but usually Float64), or a tuple the latter.

Evaluation

SpectralKit.basis_at — Function

basis_at(basis, x)

Return an iterable with known element type and length (Base.HasEltype(), Base.HasLength()) of basis functions in basis evaluated at x.

Methods are type stable.

Note

Consequences are undefined when evaluating outside the domain.

SpectralKit.linear_combination — Function

linear_combination(basis, θ, x)

Evaluate the linear combination of $∑ θₖ⋅fₖ(x)$ of function basis $f₁, …$ at x, for the given order.

The length of θ should equal dimension(θ).

linear_combination(basis, θ)

Return a callable that calculates linear_combination(basis, θ, x) when called with x.

Grids and collocation

SpectralKit.grid — Function

grid([T], basis)

Return a grid recommended for collocation, with dimension(basis) elements.

T is used as a hint for the element type of grid coordinates, and defaults to Float64. The actual type can be broadened as required. Methods are type stable.

SpectralKit.collocation_matrix — Function

collocation_matrix(basis, x)

Convenience function to obtain a collocation matrix at gridpoints x, which is assumed to have a concrete eltype.

Methods are type stable.

Internals

This section of the documentation is probably only relevant to contributors and others who want to understand the internals.

Simplified API for adding custom transformations

SpectralKit.UnivariateTransformation — Type

abstract type UnivariateTransformation

An abstract type for univariate transformations. Transformations are not required to be subtypes, this just documents the interface they need to support:

to_domain

from_domain

SpectralKit.to_domain — Function

to_domain(transformation, parent, x)

!!! FIXME document, especially differentiability requirements at infinite endpoints

SpectralKit.from_domain — Function

from_domain(transformation, parent, x)

!!! FIXME document, especially differentiability requirements at infinite endpoints

Grid internals

SpectralKit.gridpoint — Function

gridpoint(basis, i)

Return a gridpoint for collocation, with 1 ≤ i ≤ dimension(basis).

T is used as a hint for the element type of grid coordinates, and defaults to Float64. The actual type can be broadened as required. Methods are type stable.

Note

Not all grids have this method defined, especially if it is impractical. See grid, which is part of the API, this function isn't.