SpectralKit

This is a very simple package for building blocks of spectral methods. Its intended audience is users who are familiar with the theory and practice of these methods, and prefer to assemble their code from modular building blocks.

The following provides an overview of the interface for univariate function families on various domains:

julia> using SpectralKit

julia> F = ChebyshevInterval(0, 2.0)  # Chebyshev polynomial family on [0,2] — see others below
ChebyshevInterval(0.0, 2.0)

julia> is_function_family(F)          # tells us that we support the interface below
true

julia> domain_extrema(F)              # endpoints
(0.0, 2.0)

julia> roots(F, 5)                    # Gauss-Chebyshev grid
5-element Array{Float64,1}:
 0.04894348370484636
 0.412214747707527
 1.0
 1.5877852522924731
 1.9510565162951536

julia> augmented_extrema(F, 5)        # Gauss-Lobatto grid
5-element Array{Float64,1}:
 0.0
 0.2928932188134524
 1.0
 1.7071067811865475
 2.0

julia> basis_function(F, 3, 0.41, Order(0)) # value of the 3rd (starting at 1!) polynomial at 0.41
-0.3037999999999994

julia> basis_function(F, 3, 0.41, OrdersTo(1)) # value and the derivative
2-element StaticArrays.SArray{Tuple{2},Float64,1,2} with indices SOneTo(2):
 -0.3037999999999994
 -2.3600000000000008

Abstract interface for function families

SpectralKit.Order — Type

Order(D)

(Evaluate) derivative D, ≥ 0.

SpectralKit.OrdersTo — Type

OrdersTo(D)

(Evaluate) derivatives 0, …, D.

SpectralKit.is_function_family — Function

is_function_family(F)

is_function_family(f::F)

Test if the argument is a function family, supporting the following interface:

domain_extrema for querying the domain,

basis_function for function evaluation,

roots and augmented_extrema to obtain collocation points.

Can be used on both types (preferred) and values (for convenience).

SpectralKit.domain_extrema — Function

domain_extrema(family)

Return the extrema of the domain of the given family as a tuple.

Type can be arbitrary, but guaranteed to be the same for both endpoints, and type stable.

SpectralKit.basis_function — Function

basis_function(family, k, x, order)

Evaluate the kth (starting from 1) function in family at x.

order determines the derivatives:

Order(d) returns the dth derivative as a scalar, starting from 0 (for the function value)

OrdersTo(d) returns the derivatives up to d, starting from the function value, as an SVector.

The implementation is intended to be type stable.

Note

Consequences are undefined for evaluating outside the domain.

Note about indexing

Most texts index polynomial families with n = 0, 1, …. Following the Julia array indexing convention, this package uses k = 1, 2, …. Some code may use n = k - 1 internally for easier comparison with well-known formulas.

SpectralKit.linear_combination — Function

linear_combination(family, θ, x, order)

Evaluate the linear combination of $∑ θₖ⋅fₖ(x)$ of functiona family $f₁, …$ at x, for the given order.

SpectralKit.roots — Function

roots([T], family, N)

Return the roots of the K = N + 1th function in family, as a vector of N numbers with element type T (default Float64).

In the context of collocation, this is also known as the “Gauss-Chebyshev” grid.

Order is monotone, but not guaranteed to be increasing.

SpectralKit.augmented_extrema — Function

augmented_extrema([T], family, N)

Return the augmented extrema (extrema + boundary values) of the Nth function in family, as a vector of N numbers with element type T (default Float64).

In the context of collocation, this is also known as the “Gauss-Lobatto” grid.

Order is monotone, but not guaranteed to be increasing.

Specific function families

SpectralKit.Chebyshev — Type

struct Chebyshev <: SpectralKit.FunctionFamily

Chebyhev polynomials of the first kind, defined on [-1,1].

SpectralKit.ChebyshevSemiInf — Type

struct ChebyshevSemiInf{T<:Real} <: SpectralKit.TransformedChebyshev

Chebyshev polynomials transformed to the domain [A, Inf) (when L > 0) or (-Inf,A] (when L < 0) using $y = A + L ⋅ (1 + x) / (1 - x)$.

SpectralKit.ChebyshevInf — Type

struct ChebyshevInf{T<:Real} <: SpectralKit.TransformedChebyshev

Chebyshev polynomials transformed to the domain (-Inf, Inf). (when L < 0) using $y = A + L ⋅ x / √(L^2 + x^2)$.

SpectralKit.ChebyshevInterval — Type

struct ChebyshevInterval{T<:Real} <: SpectralKit.TransformedChebyshev

Chebyshev polynomials transformed to the domain (a, b). using $y = x ⋅ s + m$.

m and s are calculated and checked by the constructor; a < b is enforced.